Memory management: enhance memory cleaning infrastructure
Add a new memory cleaner type: a "waiting" cleaner, i.e. a cleaner
that can optionally wait (for work to be done by peripherals or
other CPUs) when releasing memory. Waiting cleaners are invoked
with a set of flags that specify whether they are allowed to wait
and whether their invocation is triggered by an out-of-memory
condition.
Add a mem_alloc() function, that takes a set of allocation flags,
and optionally invokes memory cleaners (via a new mem_clean()
function) if the allocation fails.
Move initialization of memory cleaning data structures to
init_kernel_heaps() so that memory cleaning API functions (and thus
mem_alloc()) can be called before kernel_runtime_init().
Change the mm_service() function to a new mem_service() function
that is meant to be called only from a background task and returns
whether memory has been cleaned during its invocation.
Change the pagecache code to call mem_clean() when page allocation
fails.
Change the anonymous page fault hander to call mem_alloc() when
allocating an anonymous page.
These changes are intended to create a generic infrastructure for
robust handling of low-memory and out-of-memory conditions.